update :amount can contain float

+++ db/migrate/20190304173437_create_recipe_ingredients.rb
@@ -3,7 +3,7 @@ class CreateRecipeIngredients < ActiveRecord::Migration[5.2]
     create_table :recipe_ingredients do |r|
       r.integer :recipe_id
       r.integer :ingredient_id
-      r.integer :amount
+      r.float :amount
     end
   end
 end
